+++ src/pyrefdev/indexer/__main__.py
@@ -3,6 +3,7 @@ import cyclopts
 import pyrefdev
 from pyrefdev.indexer.add_docs import add_docs
 from pyrefdev.indexer.crawl_docs import crawl_docs
+from pyrefdev.indexer.crawl_pypi import crawl_pypi
 from pyrefdev.indexer.parse_docs import parse_docs
 from pyrefdev.indexer.update_docs import update_docs
 from pyrefdev.indexer.update_landing_page import update_landing_page
@@ -17,6 +18,7 @@ app = cyclopts.App(
 )
 app.command(add_docs)
 app.command(crawl_docs)
+app.command(crawl_pypi)
 app.command(parse_docs)
 app.command(update_docs)
 app.command(update_landing_page)