+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Support [passlib](https://passlib.readthedocs.io/en/stable/).
 - Fixed symbols for [ipython](https://ipython.readthedocs.io/en/stable/).
 - When no symbol is specified, the lucky button now redirects to a random API reference URL.
+- Indexer: in the `add-docs` command, guess the API docs based on PyPI package information when `--index-url` is not specified.
+- Indexer: add a `crawl-pypi` command to cache the PyPI package information.
 
 ## v2025.7